+++ projeto-05/design-system/packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mamede/eslint-config
 
+## 1.0.1
+
+### Patch Changes
+
+- Add two new components - Toast and Tooltip
+
 ## 1.0.0
 
 ### Major Changes