+++ projeto-05/design-system/packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mamede/react
 
+## 1.1.0
+
+### Minor Changes
+
+- Add two new components - Toast and Tooltip
+
 ## 1.0.0
 
 ### Major Changes